back/requirements.txt
sentence-transformers>=2.2.0
simple-salesforce>=1.12.0
python-dotenv>=1.0.0
anthropic>=0.39.0
openai>=1.0.0
chromadb>=0.4.0
fastapi>=0.100.0
uvicorn>=0.22.0